Criação de sistema de bolsa de valores

+++ src/main/java/com/kuuhaku/controller/postgresql/EquipmentMarketDAO.java
@@ -85,7 +85,7 @@ public class EquipmentMarketDAO {
 
 		double average;
 		try {
-			average = (double) em.createQuery("""
+			average = (int) em.createQuery("""
 					SELECT AVG(em.price)
 					FROM EquipmentMarket em
 					WHERE em.card.card = :card
@@ -124,7 +124,7 @@ public class EquipmentMarketDAO {
 
 		double average;
 		try {
-			average = (double) em.createQuery("""
+			average = (int) em.createQuery("""
 					SELECT AVG(em.price)
 					FROM EquipmentMarket em
 					WHERE em.card.card = :card
